MonoStereo starts working

+++ dmb/modeling/flow/layers/inverse_warp_flow.py
@@ -15,11 +15,11 @@ def inverse_warp_flow(img, flow, padding_mode='zeros'):
     b, _, h, w = flow.size()
 
     # [1, H, W]  copy 0-height for w times : y coord
-    i_range = torch.arange(0, h).view(1, h, 1).expand(1, h, w).type_as(flow)
+    i_range = torch.arange(0, h).view(1, h, 1).expand(1, h, w).float()
     # [1, H, W]  copy 0-width for h times  : x coord
-    j_range = torch.arange(0, w).view(1, 1, w).expand(1, h, w).type_as(flow)
+    j_range = torch.arange(0, w).view(1, 1, w).expand(1, h, w).float()
 
-    pixel_coords = torch.stack((j_range, i_range), dim=1).type_as(flow).to(flow.device)  # [1, 2, H, W]
+    pixel_coords = torch.stack((j_range, i_range), dim=1).float().to(flow.device)  # [1, 2, H, W]
     batch_pixel_coords = pixel_coords.expand(b, 2, h, w).contiguous().view(b, 2, -1)  # [B, 2, H*W]
 
     X = batch_pixel_coords[:, 0, :] + flow[:, 0, :].contiguous().view(b, -1)  # [B, H*W]
